#41[update] サインアップ後サインインするように実装
