+++ pages/signIn.tsx
@@ -1,6 +1,6 @@
+import axios from "axios";
 import { useRouter } from "next/router";
 import { useState } from "react";
-import { signIn } from "../methods/signIn";
 const initialErrorState = {
     mailAddress: false,
     password: false,
@@ -32,12 +32,28 @@ export default function SignIn() {
         event.preventDefault();
         const { name, value } = event.target;
         setFormData({ ...formData, [name]: value });
-        console.log(formData);
     };
-
+    const signIn = async () => {
+        try {
+            const response = await axios.post(url + "/api/v1/users/signin", {
+                email: formData.mailAddress,
+                password: formData.password,
+            });
+            document.cookie = "accessToken=" + response.data.jwt;
+            router.push("/");
+        } catch (error) {
+            console.log(error.response);
+            let responseErrorStatus = Object.assign({}, initialErrorStatus);
+            if (error.response.status == 422) {
+                responseErrorStatus.mailAddressValueError = true;
+            } else if (error.response.status == 400) {
+                responseErrorStatus.passwordIncorrect = true;
+            }
+            setErrorStatus(responseErrorStatus);
+        }
+    };
     const signInCheck = async (event) => {
         event.preventDefault();
-        console.log(initialErrorState);
 
         let passwordPattern =
             /^(?=.*?[a-z])(?=.*?[A-Z])(?=.*?\d)[a-zA-Z\d]{8,100}$/;
@@ -46,8 +62,7 @@ export default function SignIn() {
             formError.password = true;
         }
         if (!formError.password && !formError.mailAddress) {
-            const error = await signIn(formData);
-            setErrorStatus(error);
+            signIn();
             // router.push("/");
         }
         setFormErrors(formError);
